+++ src/test/converter2/issues/gh2802.ts
@@ -1,0 +1,11 @@
+/**
+ * Docs
+ * @alpha
+ */
+export type AlphaOk = number | string;
+
+/**
+ * Docs2
+ * @alpha
+ */
+export type AlphaNoGo = (arg: number | string) => void;